Implement DELETE of timespans
This lets us mark timespans as rubbish.

Make Scotty handle HTTP DELETEs on "/timespans".

Implement a DELETE API.

Implement GET of rubbished timespans. This retrieves all timespans that
have been rubbished prior to a given timestamp. Parsing of this
timestamp string is also implemented.

This patch adds dependencies on time and old-locale, which are used for
the timestamps.

Signed-off-by: Alexander Berntsen <alexander@plaimi.net>
Signed-off-by: Stian Ellingsen <stian@plaimi.net>
Reviewed-by: Stian Ellingsen <stian@plaimi.net>