Generalise ParsableTime
Making 't' constrained to 'ParseTime' in our 'data ParsableTime t' was
an erroneous decision, which only leads to future headache. Instead of
being able to access the context easily, it just forces extra
constraints elsewhere (which we might have to pattern match on).

This should have been obvious when writing the patch, as the context did
not alleviate any pain with instantiating the 'ParsableTime t' ADT to
'Parsable'. The instantiation is preserved in this commit, which
eliminates the context, thereby rendering the constraint superfluous in
the first place.

Signed-off-by: Alexander Berntsen <alexander@plaimi.net>
Signed-off-by: Stian Ellingsen <stian@plaimi.net>
Reviewed-by: Stian Ellingsen <stian@plaimi.net>