Allow inserting TimespanAttributes with Timespans
When frontends make a POST request to make a timespan, they need to be
able to set attributes at the same time. Implement support for this by
treating every argument that ends with an '_' as an attribute.

The patch adds type synonyms to differentiate attribute keys and values
from specified optionals in the tests.

Signed-off-by: Alexander Berntsen <alexander@plaimi.net>
Signed-off-by: Stian Ellingsen <stian@plaimi.net>
Reviewed-by: Stian Ellingsen <stian@plaimi.net>