Serve gzipped tarball of source code at GET /src
Add a requests module for meta-requests. Add GET /src which servers a
gzipped tarball of the source code.

Generate the gzipped tarball with Setup.hs.

This stuff makes AGPL-compliance easier, since we bake it into
'cabal build'. The tarball is generated pre-build. It takes all of src,
prop and test, recursively. This is naïve, and may result in garbage
being packaged if the user is not deploying a clean tempuhs. But they
*should be*. Furthermore, we need to catch any downstream hackery (which
may include new files), which hard-coding the filenames won't be good
enough to do.

This adds a dependency on process to run tar, and on unix to create the
directory.

Signed-off-by: Alexander Berntsen <alexander@plaimi.net>
Signed-off-by: Stian Ellingsen    <stian@plaimi.net>
Reviewed-by:   Stian Ellingsen    <stian@plaimi.net>